requirements.txt
alembic==1.0.7
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-http-client==3.1.0
requests==2.21.0
sendgrid==6.0.5
six==1.12.0
SQLAlchemy==1.2.18
stripe==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
